+++ Validation/email.js
@@ -1,28 +1,17 @@
 function mailValidation(mail) {
-  const atIndex = mail.indexOf("@");
-  const dotComIndex = mail.indexOf(".com");
-  const dotItIndex = mail.indexOf(".it");
-
-  // Check if "@" exists and ".com" or ".it" exists
-  if (atIndex === -1||(dotComIndex === -1 && dotItIndex === -1)) {
-    return false;
+  if (mail.indexOf("@") == -1 || (mail.indexOf(".com") == -1 && mail.indexOf(".it") == -1)) {
+    return false
   }
 
-  // Check if "@" is not at the beginning or "@" is immediately followed by ".com" or ".it"
-  if (atIndex === 0 || atIndex + 1 === dotComIndex || atIndex + 1 === dotItIndex) {
-    return false;
+  if (mail.indexOf("@") == (mail.indexOf(".com") && mail.indexOf(".it")) - 1) {
+    return false
   }
 
-  // Check if either ".com" or ".it" is at the end of the email
-  if (dotComIndex !== -1 && dotComIndex + 4 === mail.length) {
-    return true;
+  if (mail.indexOf("@") == 0 || (mail.indexOf(".com") + 4 && mail.indexOf(".it") + 3) != mail.length) {
+    return false
   }
 
-  if (dotItIndex !== -1 && dotItIndex + 3 === mail.length) {
-    return true;
-  }
-
-  return false;
+  return true
 }
 
 module.exports = {
