--- dbApi/Schema/dailyPrev.js
@@ -1,19 +1,0 @@
-const mongoose = require('mongoose');
-
-const dataSchema = new mongoose.Schema({
-    _id: false,
-    temperatureMax: Number,
-    temperatureMin: Number
-})
-
-const dailyPrevSchema = new mongoose.Schema({
-    cityName: String,
-    countryCode: String,
-    latitude: Number,
-    longitude: Number,
-    daily: Boolean,
-    date: String,
-    data: dataSchema,
-})
-
-module.exports = {dailyPrevSchema};
