+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "mocha",
+    "test": "echo \"Error: no test specified\" && exit 1",
     "api": "nodemon ."
   },
   "author": "",
@@ -12,16 +12,9 @@
   "dependencies": {
     "axios": "^1.6.8",
     "bcryptjs": "^2.4.3",
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
-    "express": "^4.19.2",
-    "jsonwebtoken": "^9.0.2",
-    "mongoose": "^8.3.2",
-    "passport-local-mongoose": "^8.0.0"
+    "express": "^4.19.2"
   },
   "devDependencies": {
-    "chai": "^4.3.0",
-    "mocha": "^10.4.0",
     "nodemon": "^3.1.0"
   }
 }
